do not write to file on load
ini_write was used to initially load the ini data in, causing the file
to be written for each new item. this splits it up to an internal use
add_item and external use ini_write.

this also means the file data (including comments and stupid spacing)
will stay intact when only reading the ini.